+++ CHANGELOG.md
@@ -37,6 +37,10 @@ Release 5.4.0:
 - Improve support for ISO mDocs:
   - In `IssuerSignedItem`, tag `Instant` values with CBOR tag `0`, and `LocalDate` with CBOR tag `1004`
   - Fix calculating digests for `IssuerSignedItems`
+  - Fix serialization of device authentication bytes
+
+Release 5.3.3:
+ - ISO: Fix serialization of device authentication bytes
 
 Release 5.3.2:
  - ISO: Do not tag instants in CBOR with tag 1004